+++ gneiss/tests/test_sort.py
@@ -9,7 +9,7 @@ import numpy as np
 import pandas as pd
 import unittest
 from gneiss.sort import (niche_sort, mean_niche_estimator,
-                         ladderize, order_tips)
+                         ladderize, gradient_sort)
 import pandas.util.testing as pdt
 from skbio import TreeNode
 
@@ -196,23 +196,32 @@ class TestSort(unittest.TestCase):
         res = str(ladderize(tree, ascending=False))
         self.assertEqual(exp, res)
 
-    def test_order_tips(self):
+    def test_gradient_sort(self):
         # Makes sure that the tree is sorted according
         # a pre-set ordering
         tree = TreeNode.read([u'((a,b)c, ((g,h)e,f)d)r;'])
         exp = '(((g,h)e,f)d,(a,b)c)r;\n'
         x = pd.Series({'f': 3, 'g': 1, 'h': 2, 'a': 4, 'b': 5})
-        res = str(order_tips(tree, x))
+        res = str(gradient_sort(tree, x))
         self.assertEqual(exp, res)
 
-    def test_order_tips_descending(self):
+    def test_gradient_sort_descending(self):
         # Makes sure that the tree is sorted according
         # a pre-set ordering in descending order
         tree = TreeNode.read([u'((a,b)c, ((g,h)e,f)d)r;'])
         exp = '((b,a)c,(f,(h,g)e)d)r;\n'
         x = pd.Series({'f': 3, 'g': 1, 'h': 2, 'a': 4, 'b': 5})
-        res = str(order_tips(tree, x, ascending=False))
+        res = str(gradient_sort(tree, x, ascending=False))
         self.assertEqual(exp, res)
+
+    def test_gradient_sort_error(self):
+        # Makes sure that the tree is sorted according
+        # a pre-set ordering
+        tree = TreeNode.read([u'((a,b)c, ((g,h)e,f)d)r;'])
+        x = pd.Series({'f': 'x', 'g': 'y', 'h': 'z', 'a': 'u', 'b': 'dz'})
+        with self.assertRaises(ValueError):
+            gradient_sort(tree, x)
+
 
 if __name__ == '__main__':
     unittest.main()