+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Version 0.0.2
 
 ### Features
-* Added `ladderize` and `order_tips` [#29](https://github.com/biocore/gneiss/issues/29)
+* Added `ladderize` and `gradient_sort` [#29](https://github.com/biocore/gneiss/issues/29)
 
 ### Bug fixes